Request already spent time from gitlab
Time is requested only for current user. For this to work we need 2 API calls:
1. user - requestUserId gets currently authenticated user;
2. notes - requestNotes gets all notes for given merge request.
We parse notes and sum all spent time records for current user in new reducer: spentTime.
redux-thunk is used to make it simpler to chain API calls together. We should refactor old actions to use thunk as well.
Timer bug is fixed: when timer is active and we open popup, timer wasn't running.
Run it after initializing timer data.